Replace autoloader implementation
The current autoloader implementation uses `set_include_path()`, which should be avoided.

Since PHP 5.3, there's no valid reason to use `set_include_path()` in application code anymore, as:
- it adds processing overhead and additional filesystem checks to _all_ `include`/`require` statements that follow it;
- following from that, it slows down _all_ other autoloaders as well;
- it can break in environments where the include path is set at the server level and cannot be changed due to security restrictions;
- `spl_autoload_register()` can be fed a callback that does the `include` directly, avoiding any further processing.

This alternative autoloader implementation uses a direct `include` in the callback that avoids the above drawbacks and is closer to the generally accepted approach that is now used by the PHP community (usually in the form of an autoloader generated by Composer).